Fix: 실거래가타입 변경

+++ Springboot/houscore/src/main/java/com/hs/houscore/dto/BuildingInfraDTO.java
@@ -39,8 +39,8 @@ public class BuildingInfraDTO {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class RealCost{
-        private Double buy;
-        private Double longterm;
+        private Long buy;
+        private Long longterm;
         private String monthly;
     }
 }
